+++ modules/pool/lib/pool-aggregator-loader.ts
@@ -1,77 +1,57 @@
+import { Prisma, PrismaErc4626ReviewData, PrismaTokenType } from '@prisma/client';
+import { prisma } from '../../../prisma/prisma-client';
+import { HookData } from '../../../prisma/prisma-types';
 import {
-    PrismaPoolMinimal,
-    prismaPoolWithExpandedNesting,
-    PrismaPoolWithExpandedNesting,
-    HookData,
-} from '../../../prisma/prisma-types';
-import {
-    GqlBalancePoolAprItem,
-    GqlBalancePoolAprSubItem,
-    GqlPoolDynamicData,
-    GqlPoolMinimal,
-    GqlPoolUnion,
-    QueryPoolGetPoolsArgs,
-    GqlPoolAprItem,
-    GqlPoolAprItemType,
     GqlPoolAggregator,
-    LiquidityManagement,
     QueryAggregatorPoolsArgs,
+    LiquidityManagement,
 } from '../../../apps/api/gql/generated-schema';
 import _ from 'lodash';
-import { prisma } from '../../../prisma/prisma-client';
-import { Prisma, PrismaPoolAprType } from '@prisma/client';
-import { fixedNumber } from '../../view-helpers/fixed-number';
-import { ElementData, FxData, GyroData, StableData, QuantAmmWeightedData, ReclammData } from '../subgraph-mapper';
-import { ZERO_ADDRESS } from '@balancer/sdk';
+import { FxData, GyroData, StableData, QuantAmmWeightedData, ReclammData } from '../subgraph-mapper';
 import { mapHookToGqlHook } from '../../sources/transformers';
-import { mapPoolToken, enrichWithErc4626Data, mapAprItems } from './pool-gql-mapper-helper';
+import { chainToChainId } from '../../network/chain-id-to-chain';
+
+const aggregatorPrismaValidator = Prisma.validator<Prisma.PrismaPoolDefaultArgs>()({
+    include: {
+        dynamicData: true,
+        tokens: {
+            include: {
+                token: {
+                    include: {
+                        types: true,
+                    },
+                },
+            },
+        },
+        allTokens: true,
+    },
+});
+
+type AggregatorPrismaSchema = Prisma.PrismaPoolGetPayload<typeof aggregatorPrismaValidator>;
+
+const tokenWithTypes = Prisma.validator<Prisma.PrismaTokenDefaultArgs>()({
+    include: {
+        types: true,
+    },
+});
+
+type TokenWithTypes = Prisma.PrismaTokenGetPayload<typeof tokenWithTypes>;
 
 export class PoolAggregatorLoader {
     public async aggregatorPools(args: QueryAggregatorPoolsArgs): Promise<GqlPoolAggregator[]> {
-        // add limits per default
-        args.first = args.first || 1000;
-        args.skip = args.skip || 0;
-
-        const orderBy = getOrderBy(args);
-
-        const baseQuery: Prisma.PrismaPoolFindManyArgs = {
-            take: args.first || undefined,
-            skip: args.skip || undefined,
-            orderBy,
-        };
-
-        const where = args.where || {};
-
-        const allTokensFilter = [];
-        where?.tokensIn?.forEach((token) => {
-            allTokensFilter.push({
-                allTokens: {
-                    some: {
-                        token: {
-                            address: {
-                                equals: token.toLowerCase(),
-                            },
-                        },
-                    },
-                },
-            });
-        });
-
-        if (where?.tokensNotIn) {
-            allTokensFilter.push({
-                allTokens: {
-                    every: {
-                        token: {
-                            address: {
-                                notIn: where.tokensNotIn.map((t) => t.toLowerCase()) || undefined,
-                            },
-                        },
-                    },
-                },
-            });
-        }
-
-        const filterArgs: Prisma.PrismaPoolWhereInput = {
+        const where: Prisma.PrismaPoolWhereInput = {
+            id: {
+                in: args.where?.idIn?.map((id) => id.toLowerCase()) || undefined,
+            },
+            chain: {
+                in: args.where?.chainIn || undefined,
+            },
+            protocolVersion: {
+                in: args.where?.protocolVersionIn || undefined,
+            },
+            type: {
+                in: args.where?.poolTypeIn || undefined,
+            },
             dynamicData: {
                 swapEnabled: true,
                 isPaused: false,
@@ -80,58 +60,100 @@ export class PoolAggregatorLoader {
                     gt: 0.000000000001,
                 },
                 totalLiquidity: {
-                    gt: where?.minTvl || undefined,
+                    gt: args.where?.minTvl || undefined,
                 },
-            },
-            chain: {
-                in: where?.chainIn || undefined,
-                notIn: where?.chainNotIn || undefined,
-            },
-            protocolVersion: {
-                in: where?.protocolVersionIn || undefined,
-            },
-            type: {
-                in: where?.poolTypeIn || undefined,
-                notIn: where?.poolTypeNotIn || undefined,
-            },
-            createTime: {
-                gt: where?.createTime?.gt || undefined,
-                lt: where?.createTime?.lt || undefined,
             },
             NOT: {
                 categories: {
                     has: 'BLACK_LISTED',
                 },
             },
-            AND: allTokensFilter,
-            id: {
-                in: where?.idIn?.map((id) => id.toLowerCase()) || undefined,
-                notIn: where?.idNotIn?.map((id) => id.toLowerCase()) || undefined,
-            },
+            ...(args.where?.tokensIn
+                ? {
+                      AND: args.where.tokensIn.map((token) => ({
+                          allTokens: {
+                              some: {
+                                  token: {
+                                      address: {
+                                          equals: token.toLowerCase(),
+                                      },
+                                  },
+                              },
+                          },
+                      })),
+                  }
+                : {}),
         };
 
-        const query = {
-            ...baseQuery,
-            where: {
-                ...filterArgs,
-                allTokens: {
-                    some: {
-                        token: {
-                            address: filterArgs.allTokens?.some?.token?.address,
-                        },
+        // Get all the tokens for the chain - most likely scenario is client is fetching the chain information
+        console.time('dbTokens');
+        const [dbTokens, dbTypes, erc4626ReviewData] = await Promise.all([
+            prisma.prismaToken.findMany({
+                where: {
+                    ...(args.where?.chainIn ? { chain: { in: args.where?.chainIn } } : {}),
+                },
+            }),
+            prisma.prismaTokenType.findMany({
+                where: {
+                    ...(args.where?.chainIn ? { chain: { in: args.where?.chainIn } } : {}),
+                },
+            }),
+            prisma.prismaErc4626ReviewData.findMany({
+                where: {
+                    chain: {
+                        in: args.where?.chainIn || undefined,
                     },
                 },
-            },
-        };
+            }),
+        ]);
 
-        const pools = await prisma.prismaPool.findMany({
-            ...query,
-            include: {
-                ...this.getPoolInclude(),
+        const typesMap = dbTypes.reduce(
+            (agg, item) => {
+                agg[`${item.chain}-${item.tokenAddress}`] ||= [];
+                agg[`${item.chain}-${item.tokenAddress}`].push(item);
+                return agg;
             },
-        });
+            {} as Record<string, PrismaTokenType[]>,
+        );
 
-        const gqlPools = pools.map((pool) => this.mapPoolToAggregatorPool(pool));
+        const tokensMap = Object.fromEntries(
+            dbTokens.map((token) => [
+                `${token.chain}-${token.address}`,
+                { ...token, types: typesMap[`${token.chain}-${token.address}`] },
+            ]),
+        );
+        console.timeEnd('dbTokens');
+
+        console.time('dbPools');
+        const pools = await prisma.prismaPool
+            .findMany({
+                where,
+                include: {
+                    dynamicData: true,
+                    tokens: true,
+                    ...(args.where?.tokensIn
+                        ? {
+                              allTokens: true,
+                          }
+                        : {}),
+                },
+            })
+            .then((pools) =>
+                pools.map((pool) => ({
+                    ...pool,
+                    tokens: pool.tokens.map((token) => ({
+                        ...token,
+                        token: tokensMap[`${token.chain}-${token.address}`],
+                    })),
+                })),
+            );
+        console.timeEnd('dbPools');
+
+        // Get review data
+        const erc4626ReviewDataMap = Object.fromEntries(erc4626ReviewData.map((data) => [data.erc4626Address, data]));
+
+        console.time('poolsMapping');
+        const gqlPools = pools.map((pool) => this.mapPoolToAggregatorPool(pool, tokensMap, erc4626ReviewDataMap));
         const filteredPools = [];
 
         for (const mappedPool of gqlPools) {
@@ -142,141 +164,173 @@ export class PoolAggregatorLoader {
                 }
             }
 
-            // load rate provider data into PoolTokenDetail model
-            await this.enrichWithRateproviderData(mappedPool);
-
-            // load underlying token info into PoolTokenDetail
-            await enrichWithErc4626Data(mappedPool.poolTokens, mappedPool.chain);
-
             filteredPools.push(mappedPool);
         }
+        console.timeEnd('poolsMapping');
 
         return filteredPools;
     }
 
-    private async enrichWithRateproviderData(mappedPool: GqlPoolMinimal | GqlPoolAggregator | GqlPoolUnion) {
-        for (const token of mappedPool.poolTokens) {
-            if (token.priceRateProvider && token.priceRateProvider !== ZERO_ADDRESS) {
-                const rateproviderData = await prisma.prismaPriceRateProviderData.findUnique({
-                    where: {
-                        chain_rateProviderAddress: {
-                            chain: mappedPool.chain,
-                            rateProviderAddress: token.priceRateProvider,
-                        },
-                    },
-                });
-                if (rateproviderData) {
-                    token.priceRateProviderData = {
-                        ...rateproviderData,
-                        warnings: rateproviderData.warnings?.split(',') || [],
-                        upgradeableComponents:
-                            (rateproviderData.upgradableComponents as {
-                                implementationReviewed: string;
-                                entryPoint: string;
-                            }[]) || [],
-                        address: rateproviderData.rateProviderAddress,
-                        reviewFile: rateproviderData.reviewUrl,
-                    };
-                }
-            }
-            if (token.hasNestedPool) {
-                for (const nestedToken of token.nestedPool!.tokens) {
-                    if (nestedToken.priceRateProvider && nestedToken.priceRateProvider !== ZERO_ADDRESS) {
-                        const rateproviderData = await prisma.prismaPriceRateProviderData.findUnique({
-                            where: {
-                                chain_rateProviderAddress: {
-                                    chain: mappedPool.chain,
-                                    rateProviderAddress: nestedToken.priceRateProvider,
-                                },
-                            },
-                        });
-                        if (rateproviderData) {
-                            nestedToken.priceRateProviderData = {
-                                ...rateproviderData,
-                                warnings: rateproviderData.warnings?.split(',') || [],
-                                upgradeableComponents:
-                                    (rateproviderData.upgradableComponents as {
-                                        implementationReviewed: string;
-                                        entryPoint: string;
-                                    }[]) || [],
-                                address: rateproviderData.rateProviderAddress,
-                            };
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    private mapPoolToAggregatorPool(pool: PrismaPoolWithExpandedNesting): GqlPoolAggregator {
+    private mapPoolToAggregatorPool(
+        pool: AggregatorPrismaSchema,
+        underlyingMap: Record<string, TokenWithTypes>,
+        reviewMap: Record<string, PrismaErc4626ReviewData>,
+    ): GqlPoolAggregator {
         const { typeData, ...poolWithoutTypeData } = pool;
 
         const hook = (pool.hook as HookData)?.address ? (pool.hook as HookData) : null;
 
         const mappedData = {
+            ...poolWithoutTypeData,
             decimals: 18,
-            dynamicData: this.getPoolDynamicData(pool),
-            poolTokens: pool.tokens.map((token) => mapPoolToken(token, pool.protocolVersion)),
-            vaultVersion: poolWithoutTypeData.protocolVersion,
+            swapFee: pool.dynamicData!.swapFee,
+            dynamicData: {
+                ...pool.dynamicData,
+                apr: {
+                    apr: {
+                        total: '0',
+                    },
+                    thirdPartyApr: {
+                        total: '0',
+                    },
+                    nativeRewardApr: {
+                        total: '0',
+                    },
+                    swapApr: '0',
+                    hasRewardApr: false,
+                    items: [],
+                },
+                poolId: pool.dynamicData!.poolId,
+                aggregateSwapFee: pool.dynamicData!.aggregateSwapFee,
+                aggregateYieldFee: pool.dynamicData!.aggregateYieldFee,
+                aprItems: [],
+                totalSupply: pool.dynamicData!.totalShares,
+                isInRecoveryMode: pool.dynamicData!.isInRecoveryMode,
+                isPaused: pool.dynamicData!.isPaused,
+                swapEnabled: pool.dynamicData!.swapEnabled,
+                swapFee: pool.dynamicData!.swapFee,
+                totalLiquidity: `${pool.dynamicData!.totalLiquidity}`,
+                totalLiquidity24hAgo: `${pool.dynamicData!.totalLiquidity24hAgo}`,
+                totalShares: pool.dynamicData!.totalShares,
+                totalShares24hAgo: pool.dynamicData!.totalShares24hAgo,
+                swapsCount: '0',
+                lifetimeSwapFees: '0',
+                lifetimeVolume: '0',
+                protocolFees24h: '0',
+                protocolFees48h: '0',
+                surplus24h: '0',
+                surplus48h: '0',
+                volume24h: '0',
+                volume48h: '0',
+                yieldCapture24h: '0',
+                yieldCapture48h: '0',
+                protocolYieldCapture24h: '0',
+                protocolYieldCapture48h: '0',
+                volume24hAth: '0',
+                volume24hAthTimestamp: 0,
+                volume24hAtl: '0',
+                volume24hAtlTimestamp: 0,
+                fees24h: '0',
+                fees48h: '0',
+                holdersCount: '0',
+                fees24hAth: '0',
+                fees24hAthTimestamp: 0,
+                fees24hAtl: '0',
+                fees24hAtlTimestamp: 0,
+                sharePriceAth: '0',
+                sharePriceAthTimestamp: 0,
+                sharePriceAtl: '0',
+                sharePriceAtlTimestamp: 0,
+                totalLiquidityAth: '0',
+                totalLiquidityAthTimestamp: 0,
+                totalLiquidityAtl: '0',
+                totalLiquidityAtlTimestamp: 0,
+            },
+            poolTokens: pool.tokens.map((token) => {
+                const underlying = token.token.underlyingTokenAddress
+                    ? underlyingMap[`${token.chain}-${token.token.underlyingTokenAddress}`]
+                    : null;
+                const underlyingTypes = underlying?.types?.map((t) => t.type) || [];
+                const review = reviewMap[token.address] || {};
+                const types = token.token.types?.map((t) => t.type) || [];
+                return {
+                    address: token.address,
+                    name: token.token.name,
+                    symbol: token.token.symbol,
+                    decimals: token.token.decimals,
+                    balance: token.balance,
+                    weight: token.weight,
+                    isErc4626: token.token.types ? types.includes('ERC4626') : false,
+                    balanceUSD: `${token.balanceUSD}`,
+                    hasNestedPool: token.token.types ? token.address !== pool.address && types.includes('BPT') : false,
+                    index: token.index,
+                    id: token.id,
+                    isAllowed: types.includes('BLOCKED_V2') || types.includes('BLOCKED_V3'),
+                    isBufferAllowed: token.token.isBufferAllowed,
+                    isExemptFromProtocolYieldFee: token.exemptFromProtocolYieldFee,
+                    canUseBufferForSwaps: review.canUseBufferForSwaps,
+                    useUnderlyingForAddRemove: review.useUnderlyingForAddRemove,
+                    useWrappedForAddRemove: review.useWrappedForAddRemove,
+                    priceRate: token.priceRate,
+                    priceRateProvider: token.priceRateProvider,
+                    underlyingToken: underlying
+                        ? {
+                              address: underlying.address,
+                              symbol: underlying.symbol,
+                              name: underlying.name,
+                              decimals: underlying.decimals,
+                              isBufferAllowed: underlying.isBufferAllowed,
+                              chain: underlying.chain,
+                              chainId: Number(chainToChainId[underlying.chain]),
+                              isErc4626: underlyingTypes?.includes('ERC4626') || false,
+                              tradable:
+                                  underlyingTypes?.includes('BPT') || underlyingTypes?.includes('PHANTOM_BPT') || false,
+                              priority: 0,
+                          }
+                        : undefined,
+                };
+            }),
+            protocolVersion: pool.protocolVersion,
             liquidityManagement: (pool.liquidityManagement as LiquidityManagement) || undefined,
-            hook: mapHookToGqlHook(hook as HookData),
+            hook: hook ? mapHookToGqlHook(hook) : undefined,
         };
 
         switch (pool.type) {
             case 'STABLE':
-                return {
-                    ...poolWithoutTypeData,
-                    ...(typeData as StableData),
-                    ...mappedData,
-                };
             case 'META_STABLE':
-                return {
-                    ...poolWithoutTypeData,
-                    ...(typeData as StableData),
-                    ...mappedData,
-                };
             case 'COMPOSABLE_STABLE':
                 return {
-                    ...poolWithoutTypeData,
-                    ...(typeData as StableData),
                     ...mappedData,
-                    // bptPriceRate: bpt?.priceRate || '1.0',
-                };
-            case 'ELEMENT':
-                return {
-                    ...poolWithoutTypeData,
-                    ...(typeData as ElementData),
-                    ...mappedData,
-                };
-            case 'LIQUIDITY_BOOTSTRAPPING':
-                return {
-                    ...poolWithoutTypeData,
-                    ...mappedData,
+                    amp: (typeData as StableData).amp,
                 };
             case 'GYRO':
             case 'GYRO3':
             case 'GYROE':
                 return {
-                    ...poolWithoutTypeData,
-                    ...(typeData as GyroData),
                     ...mappedData,
+                    ...(typeData as GyroData), // Deprecated
                 };
             case 'FX':
                 return {
-                    ...poolWithoutTypeData,
                     ...mappedData,
-                    ...(typeData as FxData),
+                    ...(typeData as FxData), // Deprecated
                 };
             case 'QUANT_AMM_WEIGHTED':
                 return {
-                    ...poolWithoutTypeData,
                     ...mappedData,
                     quantAmmWeightedParams: typeData as QuantAmmWeightedData,
                 };
             case 'RECLAMM':
                 return {
-                    ...poolWithoutTypeData,
-                    ...(typeData as ReclammData),
+                    ...mappedData,
+                    ...(typeData as ReclammData), // Deprecated
+                };
+            case 'ELEMENT':
+                return {
+                    ...mappedData,
+                };
+            case 'LIQUIDITY_BOOTSTRAPPING':
+                return {
                     ...mappedData,
                 };
         }
@@ -286,110 +340,4 @@ export class PoolAggregatorLoader {
             ...mappedData,
         };
     }
-
-    private getPoolDynamicData(pool: PrismaPoolMinimal): GqlPoolDynamicData {
-        const {
-            fees24h,
-            totalLiquidity,
-            volume24h,
-            surplus24h,
-            fees48h,
-            volume48h,
-            surplus48h,
-            yieldCapture24h,
-            yieldCapture48h,
-            totalLiquidity24hAgo,
-            totalShares24hAgo,
-            lifetimeVolume,
-            lifetimeSwapFees,
-            holdersCount,
-            swapsCount,
-            protocolFees24h,
-            protocolFees48h,
-            protocolYieldCapture24h,
-            protocolYieldCapture48h,
-        } = pool.dynamicData!;
-
-        const newAprItemsSchema = mapAprItems(pool);
-        return {
-            ...pool.dynamicData!,
-            totalLiquidity: `${fixedNumber(totalLiquidity, 2)}`,
-            totalLiquidity24hAgo: `${fixedNumber(totalLiquidity24hAgo, 2)}`,
-            totalShares24hAgo,
-            totalSupply: pool.dynamicData?.totalShares || '0',
-            fees24h: `${fixedNumber(fees24h, 2)}`,
-            volume24h: `${fixedNumber(volume24h, 2)}`,
-            surplus24h: `${fixedNumber(surplus24h, 2)}`,
-            surplus48h: `${fixedNumber(surplus48h, 2)}`,
-            yieldCapture24h: `${fixedNumber(yieldCapture24h, 2)}`,
-            yieldCapture48h: `${fixedNumber(yieldCapture48h, 2)}`,
-            fees48h: `${fixedNumber(fees48h, 2)}`,
-            volume48h: `${fixedNumber(volume48h, 2)}`,
-            lifetimeVolume: `${fixedNumber(lifetimeVolume, 2)}`,
-            lifetimeSwapFees: `${fixedNumber(lifetimeSwapFees, 2)}`,
-            holdersCount: `${holdersCount}`,
-            swapsCount: `${swapsCount}`,
-            sharePriceAth: '0',
-            sharePriceAtl: '0',
-            totalLiquidityAth: '0',
-            totalLiquidityAtl: '0',
-            volume24hAtl: '0',
-            volume24hAth: '0',
-            fees24hAtl: '0',
-            fees24hAth: '0',
-            sharePriceAthTimestamp: 0,
-            sharePriceAtlTimestamp: 0,
-            totalLiquidityAthTimestamp: 0,
-            totalLiquidityAtlTimestamp: 0,
-            fees24hAthTimestamp: 0,
-            fees24hAtlTimestamp: 0,
-            volume24hAthTimestamp: 0,
-            volume24hAtlTimestamp: 0,
-            protocolYieldCapture24h: `${fixedNumber(protocolYieldCapture24h || 0, 2)}`,
-            protocolYieldCapture48h: `${fixedNumber(protocolYieldCapture48h || 0, 2)}`,
-            protocolFees24h: `${fixedNumber(protocolFees24h || 0, 2)}`,
-            protocolFees48h: `${fixedNumber(protocolFees48h || 0, 2)}`,
-            aprItems: newAprItemsSchema,
-            apr: {
-                apr: { __typename: 'GqlPoolAprTotal', total: '0' },
-                swapApr: '0',
-                nativeRewardApr: { __typename: 'GqlPoolAprTotal', total: '0' },
-                thirdPartyApr: { __typename: 'GqlPoolAprTotal', total: '0' },
-                items: [],
-                hasRewardApr: false,
-            },
-        };
-    }
-
-    private getPoolInclude(userAddress?: string) {
-        return {
-            ...prismaPoolWithExpandedNesting.include,
-            userWalletBalances: false,
-        };
-    }
 }
-
-const orderingColumnsMap = {
-    totalLiquidity: 'totalLiquidity',
-    totalShares: 'totalSharesNum',
-    volume24h: 'volume24h',
-    fees24h: 'fees24h',
-    apr: 'apr',
-};
-
-const getOrderBy = (args: QueryPoolGetPoolsArgs) => {
-    const orderDirection = args.orderDirection || 'desc';
-    const orderColumn = orderingColumnsMap[(args.orderBy || 'totalLiquidity') as keyof typeof orderingColumnsMap];
-
-    if (!orderColumn) {
-        return undefined;
-    }
-
-    const orderBy = {
-        dynamicData: {
-            [orderColumn]: orderDirection,
-        },
-    };
-
-    return orderBy;
-};